added 'cityAddForm' do index.html
