+++ TripPlaner/src/main/resources/templates/city.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <title>Singapore</title>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css" rel="stylesheet"
+          integrity="sha384-GLhlTQ8iRABdZLl6O3oVMWSktQOp6b7In1Zl3/Jr59b6EGGoI1aFkw7cmDA6j6gD" crossorigin="anonymous">
+</head>
+<body>
+<table class="table table-sm">
+
+    <tr th:each="city: ${cities}">
+        <td th:text="${city.id}"/>
+        <td th:text="${city.name}"/>
+        <td><a th:href="@{/city/details(id=${city.id})}" value="button"> details </a> </td>
+    </tr>
+</table>
+
+<div class="col-12">
+    <a th:href="@{city/create}">Dodaj miasto</a>
+</div>
+</body>
+</html>
