+++ TripPlaner/src/main/resources/templates/cityAddForm.html
@@ -1,26 +1,27 @@
 <!doctype html>
-<html lang="en">
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
 <head>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css" rel="stylesheet">
   <link href="https://getbootstrap.com/docs/5.3/assets/css/docs.css" rel="stylesheet">
-  <title>Singapore</title>
+  <title>Singapore - create City</title>
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js"></script>
 </head>
 <body class="p-3 m-0 border-0 bd-example">
-
-<form class="row g-3">
+<div action="#" th:action="@{/city}" th:object="${city}" method="post">
+<form class="row g-3" >
   <div class="col-md-6">
     <label for="inputCityName" class="form-label">Nazwa miasta:</label>
-    <input type="text" class="form-control" id="inputCityName" required>
+    <input type="text" class="form-control" id="inputCityName" required th:field="*{name}">
   </div>
 
   <div class="col-12">
     <label for="Textarea" class="form-label">Opis miasta:</label>
-    <textarea class="form-control" id="Textarea" rows="3" required></textarea>
+      <input type="text" class="form-control" id="Textarea" required th:field="*{description}">
   </div>
-  <div class="col-md-6">
+
+  <div class="col-md-4">
     <label for="inputCountry" class="form-label">Kraj</label>
     <input type="text" class="form-control" id="inputCountry">
   </div>
@@ -38,8 +39,9 @@
     </select>
   </div>
   <div class="col-12">
-    <button type="submit" class="btn btn-primary">Dodaj</button>
+    <button type="submit" value="Submit" class="btn btn-primary">Dodaj</button>
   </div>
 </form>
+</div>
 </body>
 </html>
