+++ TripPlaner/src/main/resources/templates/cityDetails.html
