--- TripPlaner/src/main/resources/templates/examples/ex1.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Singapore</title>
-</head>
-<body>
-<table>
-    <tr th:each="city: ${cities}">
-        <td th:text="${city.id}"/>
-        <td th:text="${city.name}"/>
-        <td><a th:href="@{/ex1/details(id=${city.id})}"> details </a> </td>
-    </tr>
-</table>
-</body>
-</html>
