+++ TripPlaner/src/main/resources/templates/index.html
@@ -136,7 +136,7 @@
                     <h5 class="card-title">Miasta</h5>
                     <p class="card-text">Polska to kraj o bogatej historii i kulturze, w którym znajdują się wiele
                         pięknych miast, z których każde ma swoje unikalne cechy i uroki. Wybierz coś dla siebie.</p>
-                    <a href="ex1" class="btn btn-primary">Zobacz</a>
+                    <a href="city" class="btn btn-primary">Zobacz</a>
                 </div>
             </div>
         </div>
